requirements.txt
streamlit
anthropic
crewai
langchain-anthropic
PyPDF2
pandas
python-dotenv